working on plotting

+++ modules/data.py
@@ -218,8 +218,9 @@ def load_data(filepath, cw_plen, cw_flen, pw_len, pw_off, seq_len, stride=1, med
     
 
     # Convert Numpy Arrays to PyTorch Tensors
-    data_new = torch.from_numpy(data_new)
-    target_new = torch.from_numpy(target_new)
+    data_new = torch.from_numpy(data_new).float()
+    target_new = torch.from_numpy(target_new).float()
+
     n = target_new.size(0) * target_new.size(1)
     mean_target_data = torch.mean(target_new.reshape(n, -1), 0)
     std_target_data = torch.std(target_new.reshape(n, -1), 0)
@@ -228,10 +229,7 @@ def load_data(filepath, cw_plen, cw_flen, pw_len, pw_off, seq_len, stride=1, med
     n=data_new.size(0) * data_new.size(1)
     mean_train_data = torch.mean(data_new.reshape(n, -1), 0)
     std_train_data = torch.std(data_new.reshape(n, -1), 0)
-
     train_data_norm = (data_new - mean_train_data) / std_train_data
-
-
 
     return train_data_norm, label_data_norm, mean_train_data, std_train_data, mean_target_data, std_target_data
     # return test_sample, data_new, target_new
